pyproject.toml
[project]
name = "tunetracker"
version = "0.1.0"
description = ""
authors = [
    { name = "Dawid Mielewczyk", email = "md.mielewczykdawid@gmail.com" },
    { name = "Wojciech Szamocki", email = "s188909@student.pg.edu.pl" },
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "pyspark (>=3.5.6,<3.6.0)",
    "kafka-python[crc32c] (>=2.2.15,<3.0.0)",
    "typer (>=0.12.3,<1.0.0)",
    "pandas (>=2.3.1,<3.0.0)",
    "numpy (>=2.3.1,<3.0.0)",
    "findspark (>=2.0.1,<3.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
]


[tool.poetry]

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"

[tool.poetry.scripts]
tunetracker = "tunetracker.main:app"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
